backend/requirements-full.txt
# =============================================================================
# Cerebrum AI Platform - Production Dependencies
# BUILD_ID: 5246938-force-rebuild
# =============================================================================

# FastAPI & Web Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0

# Redis
redis==5.0.1

# Authentication & Security
pyjwt==2.8.0
bcrypt==4.1.1
pyotp==2.9.0
qrcode==7.4.2
pillow==10.1.0
slowapi==0.1.9
cryptography==41.0.7

# Pydantic & Validation
pydantic==2.5.2
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP Client
requests==2.31.0

# AWS SDK
boto3==1.34.0

# Google API Client (Drive, OAuth)
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Logging

# Sentry
sentry-sdk[fastapi]==1.38.0

# Utilities
python-dotenv==1.0.0
typer==0.9.0
tenacity==8.2.3

# Background Tasks (Celery)
celery==5.3.4
pytest-asyncio==0.23.7

# Local development file watching
watchdog>=3.0.0

# Force rebuild: Wed Feb 19 10:15:00 UTC 2026

# ML/Vector Search
chromadb>=0.4.22  # Vector database for semantic search
sentence-transformers>=2.2.0  # For embeddings (optional fallback)

# Use pure Python alternatives
aiohttp>=3.9.0
PyPDF2>=3.0.0
python-jose>=3.3.0
psycopg2-binary>=2.9.9
structlog==24.1.0
httpx==0.27.0


alembic==1.12.1
